Graphics.UI.GLUT.Debugging

Description

This module contains a simple utility routine to report any pending GL errors.

Documentation

reportErrors :: MonadIO m => m ()

Report any pending GL errors to stderr (which is typically the console). If there are no pending errors, this routine does nothing. Note that the error flags are reset after this action, i.e. there are no pending errors left afterwards.