Graphics.UI.GLUT.Types

Description

This is a purely internal module with miscellaneous types which don't really have a good place elsewhere.

Documentation

newtype Window

An opaque identifier for a top-level window or a subwindow.

Constructors

Window CInt

data Relation

A relation between a DisplayCapability and a numeric value.

Constructors

IsEqualTo
Equal.

IsNotEqualTo
Not equal.

IsLessThan
Less than and preferring larger difference (the least is best).

IsNotGreaterThan
Less than or equal and preferring larger difference (the least is best).

IsGreaterThan
Greater than and preferring larger differences (the most is best).

IsAtLeast
Greater than or equal and preferring more instead of less. This relation is useful for allocating resources like color precision or depth buffer precision where the maximum precision is generally preferred. Contrast with IsNotLessThan relation.

IsNotLessThan
Greater than or equal but preferring less instead of more. This relation is useful for allocating resources such as stencil bits or auxillary color buffers where you would rather not over-allocate.

relationToString :: Relation -> String

data MouseButton

Mouse buttons, including a wheel

Constructors

LeftButton

MiddleButton

RightButton

WheelUp

WheelDown

AdditionalButton Int

marshalMouseButton :: MouseButton -> CInt

unmarshalMouseButton :: CInt -> MouseButton